rmp: Factor out common code
