+++ src/rmp/src/utils.cpp
@@ -1,0 +1,72 @@
+// SPDX-License-Identifier: BSD-3-Clause
+// Copyright (c) 2025-2025, The OpenROAD Authors
+
+#include "utils.h"
+
+#include <fcntl.h>
+
+#include <algorithm>
+#include <vector>
+
+#include "aig/gia/gia.h"
+#include "aig/gia/giaAig.h"
+#include "base/abc/abc.h"
+#include "base/main/main.h"
+#include "base/main/mainInt.h"
+#include "cut/abc_library_factory.h"
+#include "cut/logic_cut.h"
+#include "cut/logic_extractor.h"
+#include "db_sta/dbNetwork.hh"
+#include "db_sta/dbSta.hh"
+#include "map/scl/sclSize.h"
+#include "misc/vec/vecPtr.h"
+#include "proof/dch/dch.h"
+#include "rsz/Resizer.hh"
+#include "sta/Graph.hh"
+#include "sta/GraphDelayCalc.hh"
+#include "sta/MinMax.hh"
+#include "sta/PortDirection.hh"
+#include "sta/Search.hh"
+#include "utl/Logger.h"
+#include "utl/deleter.h"
+#include "utl/unique_name.h"
+
+namespace rmp {
+
+utl::UniquePtrWithDeleter<abc::Abc_Ntk_t> WrapUnique(abc::Abc_Ntk_t* ntk)
+{
+  return utl::UniquePtrWithDeleter<abc::Abc_Ntk_t>(ntk, &abc::Abc_NtkDelete);
+}
+
+std::vector<sta::Vertex*> GetEndpoints(sta::dbSta* sta,
+                                       rsz::Resizer* resizer,
+                                       sta::Slack slack_threshold)
+{
+  std::vector<sta::Vertex*> result;
+
+  sta::dbNetwork* network = sta->getDbNetwork();
+  for (sta::Vertex* vertex : *sta->endpoints()) {
+    sta::Pin* pin = vertex->pin();
+    sta::PortDirection* direction = network->direction(vertex->pin());
+    if (!direction->isInput()) {
+      continue;
+    }
+
+    if (resizer != nullptr) {
+      if (resizer->dontTouch(pin) || resizer->dontTouch(network->net(pin))
+          || resizer->dontTouch(network->instance(pin))) {
+        continue;
+      }
+    }
+
+    const sta::Slack slack = sta->vertexSlack(vertex, sta::MinMax::max());
+
+    if (slack < slack_threshold) {
+      result.push_back(vertex);
+    }
+  }
+
+  return result;
+}
+
+}  // namespace rmp
