+++ src/rmp/src/zero_slack_strategy.cpp
@@ -19,42 +19,12 @@
 #include "sta/StaMain.hh"
 #include "sta/Units.hh"
 #include "sta/VerilogWriter.hh"
+#include "utils.h"
 #include "utl/Logger.h"
 #include "utl/deleter.h"
 #include "utl/unique_name.h"
 
 namespace rmp {
-
-std::vector<sta::Vertex*> GetNegativeEndpoints(sta::dbSta* sta,
-                                               rsz::Resizer* resizer)
-{
-  std::vector<sta::Vertex*> result;
-
-  sta::dbNetwork* network = sta->getDbNetwork();
-  for (sta::Vertex* vertex : *sta->endpoints()) {
-    sta::Pin* pin = vertex->pin();
-    sta::PortDirection* direction = network->direction(pin);
-    if (!direction->isInput()) {
-      continue;
-    }
-
-    if (resizer != nullptr) {
-      if (resizer->dontTouch(pin) || resizer->dontTouch(network->net(pin))
-          || resizer->dontTouch(network->instance(pin))) {
-        continue;
-      }
-    }
-
-    const sta::Slack slack = sta->vertexSlack(vertex, sta::MinMax::max());
-
-    if (slack > 0.0) {
-      continue;
-    }
-    result.push_back(vertex);
-  }
-
-  return result;
-}
 
 void ZeroSlackStrategy::OptimizeDesign(sta::dbSta* sta,
                                        utl::UniqueName& name_generator,
@@ -69,7 +39,7 @@ void ZeroSlackStrategy::OptimizeDesign(sta::dbSta* sta,
   sta::dbNetwork* network = sta->getDbNetwork();
 
   std::vector<sta::Vertex*> candidate_vertices
-      = GetNegativeEndpoints(sta, resizer);
+      = GetEndpoints(sta, resizer, 0.0);
 
   if (candidate_vertices.empty()) {
     logger->info(utl::RMP,
